backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Cors==4.0.0
Werkzeug==3.0.0
PyJWT==2.8.0
gunicorn==21.2.0
